+++ src/main/scala/it/pagopa/interop/partymanagement/api/impl/PartyApiMarshallerImpl.scala
@@ -51,7 +51,4 @@ object PartyApiMarshallerImpl extends PartyApiMarshaller with SprayJsonSupport w
 
   override implicit def toEntityMarshallerAttributearray: ToEntityMarshaller[Seq[Attribute]] =
     sprayJsonMarshaller[Seq[Attribute]]
-
-  override implicit def toEntityMarshallerInstitutionId: ToEntityMarshaller[InstitutionId] =
-    sprayJsonMarshaller[InstitutionId]
 }
