+++ src/main/scala/it/pagopa/interop/partymanagement/api/impl/package.scala
@@ -17,7 +17,7 @@ package object impl extends SprayJsonSupport with DefaultJsonProtocol {
   implicit val personFormat: RootJsonFormat[Person]                                   = jsonFormat1(Person)
   implicit val institutionSeedFormat: RootJsonFormat[InstitutionSeed]                 = jsonFormat11(InstitutionSeed)
   implicit val institutionFormat: RootJsonFormat[Institution]                         = jsonFormat12(Institution)
-  implicit val institutionIdFormat: RootJsonFormat[InstitutionId]                     = jsonFormat1(InstitutionId)
+  implicit val institutionIdFormat: RootJsonFormat[InstitutionId]                     = jsonFormat2(InstitutionId)
   implicit val relationshipProductFormat: RootJsonFormat[RelationshipProduct]         = jsonFormat3(RelationshipProduct)
   implicit val relationshipFormat: RootJsonFormat[Relationship]                       = jsonFormat15(Relationship)
   implicit val relationshipProductSeedFormat: RootJsonFormat[RelationshipProductSeed] = jsonFormat2(
